Replace AbstractGameScene with GameLevelScene
