+++ Nappou-2/src/main/java/org/atoiks/games/nappou2/levels/level1/Data.java
@@ -22,7 +22,7 @@ import org.atoiks.games.nappou2.entities.Message;
 
 import static org.atoiks.games.nappou2.entities.Message.HorizontalAlignment;
 
-import static org.atoiks.games.nappou2.scenes.RefittedGameScene.GAME_BORDER;
+import static org.atoiks.games.nappou2.scenes.GameLevelScene.GAME_BORDER;
 
 /* package */ final class Data {
