+++ Nappou-2/src/main/java/org/atoiks/games/nappou2/levels/level1/Easy.java
@@ -48,8 +48,8 @@ import static org.atoiks.games.nappou2.Utils.circularPathEnemy;
 
 import static org.atoiks.games.nappou2.levels.level1.Data.*;
 
-import static org.atoiks.games.nappou2.scenes.RefittedGameScene.HEIGHT;
-import static org.atoiks.games.nappou2.scenes.RefittedGameScene.GAME_BORDER;
+import static org.atoiks.games.nappou2.scenes.GameLevelScene.HEIGHT;
+import static org.atoiks.games.nappou2.scenes.GameLevelScene.GAME_BORDER;
 
 public final class Easy implements ILevelState {
