+++ Nappou-2/src/main/java/org/atoiks/games/nappou2/levels/level1/Normal.java
@@ -49,8 +49,8 @@ import static org.atoiks.games.nappou2.Utils.tweenRadialGroupPattern;
 
 import static org.atoiks.games.nappou2.levels.level1.Data.*;
 
-import static org.atoiks.games.nappou2.scenes.RefittedGameScene.HEIGHT;
-import static org.atoiks.games.nappou2.scenes.RefittedGameScene.GAME_BORDER;
+import static org.atoiks.games.nappou2.scenes.GameLevelScene.HEIGHT;
+import static org.atoiks.games.nappou2.scenes.GameLevelScene.GAME_BORDER;
 
 public final class Normal implements ILevelState {
