+++ Nappou-2/src/main/java/org/atoiks/games/nappou2/levels/level2/Easy.java
@@ -39,8 +39,8 @@ import static org.atoiks.games.nappou2.Utils.leapEnemy;
 import static org.atoiks.games.nappou2.Utils.starShotEnemy;
 import static org.atoiks.games.nappou2.Utils.singleShotEnemy;
 
-import static org.atoiks.games.nappou2.scenes.RefittedGameScene.HEIGHT;
-import static org.atoiks.games.nappou2.scenes.RefittedGameScene.GAME_BORDER;
+import static org.atoiks.games.nappou2.scenes.GameLevelScene.HEIGHT;
+import static org.atoiks.games.nappou2.scenes.GameLevelScene.GAME_BORDER;
 
 public final class Easy implements ILevelState {
