+++ Nappou-2/src/main/java/org/atoiks/games/nappou2/scenes/LoadingScene.java
@@ -54,8 +54,8 @@ import org.atoiks.games.nappou2.GameConfig;
 import org.atoiks.games.nappou2.SoundEffect;
 
 import static org.atoiks.games.nappou2.App.SANS_FONT;
-import static org.atoiks.games.nappou2.scenes.RefittedGameScene.WIDTH;
-import static org.atoiks.games.nappou2.scenes.RefittedGameScene.HEIGHT;
+import static org.atoiks.games.nappou2.scenes.GameLevelScene.WIDTH;
+import static org.atoiks.games.nappou2.scenes.GameLevelScene.HEIGHT;
 
 public final class LoadingScene implements Scene {
