+++ Nappou-2/src/main/java/org/atoiks/games/nappou2/scenes/PauseOverlay.java
@@ -26,8 +26,8 @@ import org.atoiks.games.framework2d.Input;
 import org.atoiks.games.framework2d.IGraphics;
 import org.atoiks.games.framework2d.SceneManager;
 
-import static org.atoiks.games.nappou2.scenes.AbstractGameScene.HEIGHT;
-import static org.atoiks.games.nappou2.scenes.AbstractGameScene.GAME_BORDER;
+import static org.atoiks.games.nappou2.scenes.GameLevelScene.HEIGHT;
+import static org.atoiks.games.nappou2.scenes.GameLevelScene.GAME_BORDER;
 
 /* package */ final class PauseOverlay {
