+++ Nappou-2/src/main/java/org/atoiks/games/nappou2/scenes/ScoreScene.java
@@ -33,8 +33,8 @@ import org.atoiks.games.nappou2.Difficulty;
 import org.atoiks.games.nappou2.GameConfig;
 
 import static org.atoiks.games.nappou2.App.SANS_FONT;
-import static org.atoiks.games.nappou2.scenes.RefittedGameScene.WIDTH;
-import static org.atoiks.games.nappou2.scenes.RefittedGameScene.HEIGHT;
+import static org.atoiks.games.nappou2.scenes.GameLevelScene.WIDTH;
+import static org.atoiks.games.nappou2.scenes.GameLevelScene.HEIGHT;
 
 public final class ScoreScene extends CenteringScene {
